Moved a few things around
