+++ app/src/main/java/com/sebastianvm/musicplayer/designsystem/icons/Album.kt
@@ -1,8 +1,8 @@
-package com.sebastianvm.musicplayer.ui.icons
+package com.sebastianvm.musicplayer.designsystem.icons
 
 import com.sebastianvm.musicplayer.R
 import com.sebastianvm.musicplayer.ui.util.compose.IconState
 import com.sebastianvm.musicplayer.ui.util.compose.ResourceIcon
 
-val Icons.Genre: IconState
-    get() = ResourceIcon(resId = R.drawable.ic_genre)
+val Icons.Album: IconState
+    get() = ResourceIcon(R.drawable.ic_album)
