+++ app/src/main/java/com/sebastianvm/musicplayer/designsystem/icons/Icons.kt
@@ -1,4 +1,4 @@
-package com.sebastianvm.musicplayer.ui.icons
+package com.sebastianvm.musicplayer.designsystem.icons
 
 // TODO move remaining icons to Icons extensions
 object Icons
